% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labb3.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra function.}
\usage{
dijkstra(data, init_node)
}
\arguments{
\item{data}{A data frame.}

\item{init_node}{A number.}
}
\value{
Calculate the shortest path from the initial node \code{init_node} to every other node in data frame \code{data}.
}
\description{
Dijkstra function.
}
