% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labb3.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean function.}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
The greatest common divisor of  \code{a} and \code{b}.
}
\description{
Euclidean function.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
export
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
