\name{QQQ-package}
\alias{QQQ-package}
\alias{QQQ}
\docType{package}
\title{
\packageTitle{QQQ}
}
\description{
\packageDescription{QQQ}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{QQQ}
\packageIndices{QQQ}
~~ An overview of how to use the package, including ~~
~~ the most important functions ~~
}
\author{
\packageAuthor{QQQ}

Maintainer: \packageMaintainer{QQQ}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
